feat(schema): alinear modelo de datos con documento de disenio etapa 1
Reescritura completa del schema para cumplir exactamente con el modelo
propuesto en el documento de TFI-2026 (Etapa 1: modelado de la base).

Nuevas tablas:
- estado_vehiculo: catalogo de estados (reemplaza enum embebido)
- tipo_reserva: catalogo de tipos con requiere_garantia y antelacion_max_dias
- ubicacion_vehiculo: historial de ubicacion fisica (pertenencia != ubicacion)
- historial_estado_vehiculo: historial temporal de transiciones de estado
- garantia_reserva: datos de tarjeta (hash) asociados a una reserva

Tablas modificadas:
- vehiculo: id_sucursal renombrado a id_sucursal_origen, drop estado VARCHAR,
  agregado id_estado FK a estado_vehiculo
- reserva: agregado id_tipo_reserva FK, estado IN (pendiente, concretada, cancelada)
- alquiler: agregado id_sucursal_devolucion FK, estado IN (activo, cerrado)
- factura: agregados snapshots precio_por_dia_aplicado y porcentaje_recargo_aplicado,
  fecha_emision pasa a DATE
- mantenimiento: fecha_envio y fecha_devolucion pasan a DATE

Constraints:
- Indices unicos parciales en ubicacion_vehiculo y historial_estado_vehiculo
  para garantizar un solo registro vigente por vehiculo (fecha_hasta IS NULL /
  fecha_fin IS NULL).

Triggers/funciones:
- fn_check_vehiculo_overlap reescrito para nuevo dominio de estados
- fn_alquiler_lifecycle: 3 triggers (BEFORE UPDATE marca cerrado, AFTER INSERT
  propaga alquilado y concreta reserva, AFTER UPDATE propaga disponible,
  cierra ubicacion y abre nueva en sucursal de devolucion, actualiza km)
- fn_mantenimiento_lifecycle: 2 triggers (envio propaga en_mantenimiento,
  devolucion propaga disponible)
- Guard en fn_alquiler_close: id_sucursal_devolucion obligatorio al cerrar

Seeds reescritos para cubrir las 17 tablas, con vehiculos en estado mixto
(disponible, alquilado, en_mantenimiento) y alquileres cerrados que ejercitan
los triggers de ciclo de vida via INSERT + UPDATE.

+++ schema/01_tables/09_imagen_vehiculo.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS imagen_vehiculo (
     id_imagen    BIGSERIAL PRIMARY KEY,
     id_vehiculo  BIGINT       NOT NULL,
     url_imagen   VARCHAR(500) NOT NULL,
-    orden        SMALLINT     NOT NULL,
+    orden        INTEGER      NOT NULL,
     CONSTRAINT chk_imagen_orden CHECK (orden BETWEEN 1 AND 5)
 );
